fix int to double

+++ src/lk_func.h
@@ -7,12 +7,12 @@
 #define LK_FUNC_H
 
 int** getArrBright(QImage image);
-int* calcOptFlow(subSize window, int** arrayGray, int** arrayGrayNext);
-int* matrixVectorMultiplic(int** array, int* vector);
+double* calcOptFlow(subSize window, int** arrayGray, int** arrayGrayNext);
+double* matrixVectorMultiplic(double **array, int* vector);
 void calcGrid(QImage image);
-void freeMemoryFloat(int** trash, int size);
+void freeMemoryFloat(double** trash, int size);
 void freeMemoryInt(int** trash, int size);
 void imageInfo(QImage image, int** arr);
-void inversion(int **A, int N);
+void inversion(double **A, int N);
 
 #endif
